add data to database and connected the 2

+++ ToDoApi/ToDoApi/Model/TodoItem.cs
@@ -10,6 +10,6 @@ namespace ToDoApi.Model
         public long Id { get; set; }
         public string Name { get; set; }
         public bool IsComplete { get; set; }
-        public long ListId { get; set; }
+        public int ListId { get; set; }
     }
 }
